Get rid of closure transformation
So this still needs to be cleaned up and documented better, but it's
functional. Also it shouldn't be merged because it depends on a
snapshot of Declarative. =)

In Declarative, we ran into nasty problems with complex closure
"translation" into model objects, like was being used here for
argument processing. So with Declarative 1.2, I changed to actually
transform the AST into the runtime model, using the compile-time model
(`ModelAST*` classes, from `pipeline-model-api`) and the original AST
to do so. It was a bit of work and still needs to be documented
better, but it's less mysterious and far more reliable - no more need
for weird whitelisting tricks, intermediary transformer classes in CPS
code, etc.

Note that once this gets merged, any existing Pipelines using this
plugin that are currently running will barf out upon restart. New runs
will be fine, but if a run started before upgrading this plugin and
then tried to resume after the restart, it'll fail. Probably not an
issue here, but thought I should mention it. =)

Also, @jstrachan, I'm 99.99% sure that I could/should be doing the AST
transformation stuff in better ways. This *works* but I have a feeling
the guy who actually wrote all the classes I'm calling in the first
place may have some suggestions. =)